Adding in check that work is approved for update and destroy events.
This is to avoid messages being generated for works that are being
updated prior to approval. Equally we do not want to generate
destroy messages for works that have not been previously published to
the RDSS. When considering workflows in the future we should be careful
not to have works deactivated before a destroy message is generated.